feat: initialize application routing with protected authentication flow and error boundaries
