--- src/renderer/src/e.tsx
@@ -1,11 +1,0 @@
-import './assets/main.css'
-
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import App from './App'
-
-createRoot(document.getElementById('root')!).render(
-  <StrictMode>
-    <App />
-  </StrictMode>
-)
